+++ pagination.php
@@ -1,10 +1,14 @@
 <?php
 namespace Grav\Plugin;
 
+use Composer\Autoload\ClassLoader;
 use Grav\Common\Page\Collection;
-use Grav\Common\Page\Page;
+use Grav\Common\Page\Interfaces\PageInterface;
 use Grav\Common\Plugin;
+use Grav\Plugin\Pagination\PaginationHelper;
+use Grav\Plugin\Pagination\PaginationPage;
 use RocketTheme\Toolbox\Event\Event;
+use Twig\TwigFunction;
 
 class PaginationPlugin extends Plugin
 {
@@ -19,8 +23,21 @@ class PaginationPlugin extends Plugin
     public static function getSubscribedEvents()
     {
         return [
-            'onPluginsInitialized' => ['onPluginsInitialized', 0]
+            'onPluginsInitialized' => [
+                ['autoload', 100001],
+                ['onPluginsInitialized', 0]
+            ]
         ];
+    }
+
+    /**
+     * [onPluginsInitialized:100000] Composer autoload.
+     *
+     * @return ClassLoader
+     */
+    public function autoload()
+    {
+        return require __DIR__ . '/vendor/autoload.php';
     }
 
     /**
@@ -32,6 +49,9 @@ class PaginationPlugin extends Plugin
             $this->active = false;
             return;
         }
+
+        class_alias(PaginationHelper::class, 'Grav\\Plugin\\PaginationHelper');
+        class_alias(PaginationPage::class, 'Grav\\Plugin\\PaginationPage');
 
         $this->enable([
             'onTwigTemplatePaths' => ['onTwigTemplatePaths', 0],
@@ -53,9 +73,10 @@ class PaginationPlugin extends Plugin
      */
     public function onTwigExtensions()
     {
-        require_once(__DIR__.'/twig/PaginationTwigExtension.php');
-
-        $this->grav['twig']->twig->addExtension(new PaginationTwigExtension());
+         // Add Twig functions
+        $this->grav['twig']->twig()->addFunction(
+            new TwigFunction('paginate', [$this, 'paginateTwigFunction'])
+        );
     }
 
     /**
@@ -63,7 +84,7 @@ class PaginationPlugin extends Plugin
      */
     public function onPageInitialized()
     {
-        /** @var Page $page */
+        /** @var PageInterface $page */
         $page = $this->grav['page'];
 
         if ($page && ($page->value('header.content.pagination') || $page->value('header.pagination'))) {
@@ -96,7 +117,6 @@ class PaginationPlugin extends Plugin
         }
 
         if (!empty($params['limit']) && $collection->count() > $params['limit']) {
-            require_once __DIR__ . '/classes/paginationhelper.php';
             $this->pagination = new PaginationHelper($collection);
             $collection->setParams(['pagination' => $this->pagination]);
         }
@@ -119,14 +139,13 @@ class PaginationPlugin extends Plugin
      * @param int $limit
      * @param array $ignore_param_array      url parameters to be ignored in page links
      */
-    public function paginateCollection( $collection, $limit, $ignore_param_array = [])
+    public function paginateCollection($collection, $limit, $ignore_param_array = [])
     {
         $collection->setParams(['pagination' => 'true']);
         $collection->setParams(['limit' => $limit]);
         $collection->setParams(['ignore_params' => $ignore_param_array]);
 
         if ($collection->count() > $limit) {
-            require_once __DIR__ . '/classes/paginationhelper.php';
             $this->pagination = new PaginationHelper($collection);
             $collection->setParams(['pagination' => $this->pagination]);
 
@@ -137,5 +156,10 @@ class PaginationPlugin extends Plugin
                 $collection->slice($start, $limit);
             }
         }
+    }
+
+    public function paginateTwigFunction($collection, $limit, $ignore_url_param_array = [])
+    {
+        $this->paginateCollection($collection, $limit, $ignore_url_param_array);
     }
 }