+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v1.4.3
+## 08/26/2022
+
+1. [](#new)
+    * Pass phpstan level 1 tests
+    * Require Grav v1.6
+    * Moved pagination classes under `Grav\Plugin\Pagination` namespace
+    * Removed extension to bring Twig function into Plugin class
+
 # v1.4.2
 ## 05/09/2019